Refactor Main.js page, remove Welcome.js component and attach the elements directly
