--- lpf/src/components/Welcome/Welcome.module.css
@@ -1,95 +1,0 @@
-h1, h2{
-    font-family: 'Lato', sans-serif;
-    font-family: 'Source Sans Pro', sans-serif;
-}
-
-.welcome{
-    position: absolute;
-    left: 5%;
-    top: 30%;
-    padding: 10px;
-    animation: fadeIn 2s linear forwards;
-}
-
-.welcome__message{
-    font-size: 48px;
-}
-
-.welcome__about-me-button{
-    border: 1px solid #FFFFFF;
-    color: #FFFFFF;
-    text-decoration: none;
-    background-color: transparent;
-    padding: 5px;
-    border-radius: 5px;
-    transition: background 0.5s, color 0.5s;
-}
-
-.welcome__about-me-button:hover{
-    background-color: #FFFFFF;
-    color: #666967;
-    cursor: pointer;
-    transition: background 0.5s, color 0.5s;
-}
-
-
-.profile-pic{
-    position: absolute;
-    bottom: 0%;
-    right: 15%;
-    height: 700px;
-    width: 500px;
-    animation: fadeIn 2s linear forwards;
-}
-
-.linkedIn{
-    color: #FFFFFF;
-    font-size: 32px;
-    position: absolute;
-    bottom: 5px;
-    left: 5px;
-}
-
-@media screen and (max-width: 1024px){
-   .welcome{
-       top: 150px;
-       z-index: 1;
-   }
-
-   .profile-pic{
-        height: 800px;
-        width: 600px;
-        right: 10%;
-   }
-
-   .linkedIn{
-        font-size: 48px;
-   }
-}
-
-@media screen and (max-width: 767px){
-
-   .welcome{
-       top: 140px;
-       z-index: 1;
-   }
-
-   .welcome__message{
-        font-size: 24px;
-   }
-
-   .profile-pic{
-        right: 0%;
-        height: 600px;
-        width: 400px;
-    }
-}
-
-@keyframes fadeIn{
-    0%{
-        opacity: 0%;
-    }
-    100%{
-        opacity: 100%;
-    }
-}
